feat: list games loading page
