+++ src/app/(protected)/games/loading.tsx
@@ -3,12 +3,8 @@ import { DashboardPageHeader } from '@/components/dashboard/dashboard-page-heade
 import { Navbar } from '@/components/navbar';
 import { PublicGameRoomCard } from '@/components/public-game-room-card';
 import { Separator } from '@/components/ui/separator';
-import { createCaller } from '@/utils/caller';
 
-export default async function ListGamesPage() {
-  const caller = createCaller();
-  const publicRooms = await caller.room.list();
-
+export default function ListGamesLoadingScreen() {
   return (
     <div className="flex h-full min-h-screen w-screen flex-col items-center">
       <Navbar />
@@ -23,9 +19,10 @@ export default async function ListGamesPage() {
         <Separator className="w-full" />
 
         <section className="grid grid-cols-1 gap-8 md:grid-cols-2 lg:grid-cols-3">
-          {publicRooms.map(room => (
-            <PublicGameRoomCard key={room.id} {...room} />
-          ))}
+          <PublicGameRoomCard.Skeleton />
+          <PublicGameRoomCard.Skeleton />
+          <PublicGameRoomCard.Skeleton />
+          <PublicGameRoomCard.Skeleton />
         </section>
       </main>
     </div>
